students/Khisametdinova_Dinara/Lab_3/psychologists_service/requirements.txt
alembic==1.15.1
annotated-types==0.7.0   
anyio==4.8.0
babel==2.17.0
backrefs==5.8
bcrypt==4.3.0
certifi==2025.1.31       
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2     
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0   
fastapi==0.115.11        
fastapi-cli==0.0.7       
ghp-import==2.1.0        
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.7
mkdocs-material-extensions==1.3.1
orjson==3.10.15
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pymdown-extensions==10.14.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==15.0.1
